separate top_n_heap file
